refactor(types): move ConnectorCategory enum to types/enum
